requirements.txt
selenium==4.6.0
pytest==7.2.2
faker==18.3.1
allure-pytest==2.13.1
pytest-xdist==3.2.1
